PR review comments resolved...
