+++ stories/ContentHeader/ContentHeader.stories.tsx
@@ -4,8 +4,22 @@ import {
   Props
 } from "../../src/components/Layout/ContentHeader/ContentHeader";
 import ContentHeaderDocumentation from "./ContentHeader.mdx";
-import React, { Fragment, ReactElement } from "react";
+import React, { Fragment, ReactElement, useState } from "react";
 import { Button, ButtonProps, Icon, Icons } from "../../src";
+
+import {
+  FormStepper,
+  FormStepProps,
+  Props as FormStepperProps
+} from "../../src/components/Layout/FormPage/FormWithStepper/FormStepper/FormStepper";
+import {
+  FormWithStepper,
+  Props as FormWithStepperProps
+} from "../../src/components/Layout/FormPage/FormWithStepper/FormWithStepper";
+import { FormSection } from "../../src";
+import { FormControlExample } from "../Form/Examples/FormControlExample";
+import { StepStatus } from "../../src/components/Stepper/Step";
+import { useFullHeightCollapse } from "../../src/hooks/useFullHeightCollapse";
 
 const contentButtonList: ReactElement<ButtonProps, typeof Button>[] = [
   <Button key="1" onClick={() => alert("Cancel button clicked.")} variant="text">
@@ -62,15 +76,100 @@ ContentHeaderWithSubtitle.args = {
   buttons: contentButtonList.slice(0, 2)
 };
 
-export const ContentHeaderWithCollapseHeader = Template.bind({});
+//--- ContentHeader Collapse Example ----//
+
+const initialStepperState: FormStepProps[] = [
+  {
+    status: "current",
+    label: "Step 1",
+    targetScrollViewId: "samlBaseInfo"
+  },
+  {
+    status: "waiting",
+    label: "Step 2",
+    targetScrollViewId: "samlConnectionDetail"
+  },
+  {
+    status: "done",
+    label: "Step 3",
+    targetScrollViewId: "samlFederationRequest"
+  },
+  {
+    status: "error",
+    label: "Step 4",
+    targetScrollViewId: "samlAttributeMapping"
+  }
+];
+
+const formStepperProps: FormStepperProps = {
+  direction: "vertical",
+  textPosition: "bottom",
+  steps: initialStepperState
+};
+
+const Template1: StoryFn<FormWithStepperProps> = args => {
+  
+  const [steps, setSteps] = useState(initialStepperState);
+
+  const updateStepStatus = (index: number, status: StepStatus) => {
+    setSteps(prevState => {
+      const props = prevState.concat();
+      props[index].status = status;
+      return props;
+    });
+  };
+
+  const formStepperState = { ...formStepperProps, steps: steps };
+
+  const formStepperComponent = <FormStepper {...formStepperState}></FormStepper>;
+  const { collapsed, handleScroll } = useFullHeightCollapse(false);
+  return (
+    <>
+    <ContentHeader title="Form Header" collapsed={collapsed}>
+    <p>Form Header Description</p>
+  </ContentHeader>
+    <FormWithStepper stepper={formStepperComponent} onScroll={handleScroll}>
+      <div>
+        <FormSection
+          id="samlBaseInfo"
+          header="Step 1 Form Section"
+          subtext="Form Section sub-header."
+        >
+          <FormControlExample index={0} updateStepStatus={updateStepStatus}></FormControlExample>
+        </FormSection>
+        <FormSection
+          id="samlConnectionDetail"
+          header="Step 2 Form Section"
+          subtext="Form Section Subheader"
+        >
+          <FormControlExample index={1} updateStepStatus={updateStepStatus}></FormControlExample>
+        </FormSection>
+        <FormSection
+          id="samlFederationRequest"
+          header="Step 3 Form section"
+          subtext="Form Section Subheader"
+        >
+          <FormControlExample index={2} updateStepStatus={updateStepStatus}></FormControlExample>
+        </FormSection>
+        <FormSection
+          id="samlAttributeMapping"
+          header="Step 4 Form section"
+          subtext="Form Section Subheader"
+        >
+          <FormControlExample index={3} updateStepStatus={updateStepStatus}></FormControlExample>
+        </FormSection>
+      </div>
+    </FormWithStepper>
+    </>
+  );
+};
+
+export const ContentHeaderWithCollapseHeader = Template1.bind({});
 
 ContentHeaderWithCollapseHeader.args = {
   children: (
-    <Fragment>
-      Content <i>Description</i>
-    </Fragment>
+    <p>Form Header Description</p>
   ),
-  buttons: contentButtonList,
-  collapsed: true,
-  subtitle: "Hey there! I'm just a subtitle"
+  buttons: [],
+  collapsed: false
 };
