--- stories/hooks/useContentHeaderCollapse.mdx
@@ -1,40 +1,0 @@
-import { Canvas, Story, Title, Subtitle, ArgTypes, Meta } from "@storybook/blocks";
-import * as useContentHeaderCollapse from "./useContentHeaderCollapse.stories.tsx";
-
-<Meta of={useContentHeaderCollapse} />
-<Title />
-<Subtitle />
-
-The `useContentHeaderCollapse` hook will be used for collapse and expend the ContentHeader on behalf of the scrollbar. You should use this hook whenever you want to collapse/expand a `ContentHeader`.
-
-# Example
-
-<Canvas>
-  <Story of={useContentHeaderCollapse.FormWithStepperComponentAndExpandOrCollapsibleHeader} />
-</Canvas>
-
-Example `ContentHeaderWithCollapseandExpend`:
-
-```JSX
-
-const Template: StoryFn<Props> = args => {
- 
-  const formStepperState = { ...formStepperProps, steps: steps };
-
-  const formStepperComponent = <FormStepper {...formStepperState}></FormStepper>;
-  const { collapsed, handleScroll } = useContentHeaderCollapse(false);
-  return (
-    <>
-    <ContentHeader title="Form Header" collapsed={collapsed}>
-    <p>Form Header Description</p>
-  </ContentHeader>
-  
-    <FormWithStepper stepper={formStepperComponent} onScroll={handleScroll}>
-      <div>
-        Sample Form
-      </div>
-    </FormWithStepper>
-    </>
-  );
-};
-```
